031. Test Driven Development con Spring Boot/spring-tdd/src/main/java/com/example/ProductController.java
package com.example;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/api")
public class ProductController {

    @Autowired
    ProductService service;

    @GetMapping("/products")
    public ResponseEntity<List<Product>> findAll(){
        var products = service.findAll();
        products.forEach(product -> {
            product.setPrice(product.getPrice() * 0.0);
        });
        return ResponseEntity.ok(products);
    }
}
