031. Test Driven Development con Spring Boot/spring-tdd/src/test/java/com/example/ProductControllerTest.java
package com.example;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.util.List;

import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
import static org.hamcrest.Matchers.*;

@WebMvcTest(ProductController.class)
public class ProductControllerTest {

    @Autowired
    MockMvc mvc;

    @Autowired
    ObjectMapper mapper;

    @MockBean
    ProductService service;

    @Test
    void findAll() throws Exception {
        when(service.findAll()).thenReturn(List.of(
                new Product(1L, "Product 1", 14.99),
                new Product(2L, "Product 2", 100.0)
        ));
        mvc.perform(get("/api/products").contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk()) // 200
                .andExpect(jsonPath("$", hasSize(2)))
                .andExpect(jsonPath("$[0].title", is("Product 1")))
                .andExpect(jsonPath("$[0].price", is(0.0))); // ejemplo de modificación realizada por el controlador
    }

    @Test
    void findByIdFound() {


    }
    @Test
    void findByIdNotFound() {
    }

    @Test
    void findStockStats() {
    }

    @Test
    void create() {
    }

    @Test
    void update() {
    }

    @Test
    void deleteById() {
    }
}
